Fix layout issue in Settings screen
The footer in the Security section would overlap when in landscape for
the iPhone X. This would only happen in the device and not the
simulator. I fix it by removing the dot at the end of the sentence.
Xcode also applied some auto corrections to the Storyboard possibily due
to previous merging conflicts.