Fix the compilation error for external developers using the ExternalDeveloper.xcconfig configuration.
The PR #633 added the 'Open in DuckDuckGo' action.
However the PRODUCT_BUNDLE_IDENTIFIER for the OpenAction target is hardcoded to 'com.duckduckgo.mobile.ios.OpenAction'. See https://github.com/duckduckgo/iOS/pull/633/commits/42804d85c37f8c14ed3203ea0c96f10e365f4013#diff-9e6bc0fba251950ba65ffa409e75b9b0R4792.

This is causing a build error for developers that are not part of the DuckDuckGo team and using the `ExternalDeveloper.xcconfig` configuration.

The proposed solution consists of changing the PRODUCT_BUNDLE_IDENTIFIER for the OpenAction target from the hardcoded com.duckduckgo.mobile.ios.OpenAction to "$(APP_ID).OpenAction".

This fixes the broken build for external developers and should not break the build for the DuckDuckGo team. This is also consistent with the product identifier of the other targets.