Avoid throwing exceptions when `fetch` is called with no arguments (#708)
* Return via the built-in fetch function if no arguments are provided.

Some websites will call the `fetch` function to test for browser support. This returns a rejected Promise if no arguments are provided and an exception otherwise, providing a way to use try/catch blocks to check for compatibility.

DDG’s overridden fetch function will throw an exception if no arguments are provided (whereas the built-in will not), giving a false negative. To fix this, the original fetch should be used when the arguments list is empty.

* Use threequals for checking the fetch arguments length.