Add SwiftUI previews for widgets (#711)
* Split the unit tests scheme out of the main app scheme.

This is so that the DuckDuckGo build action does not build unit tests during regular builds. This has been fine so far, but appears to trip up SwiftUI previews.

* Add dark and light mode previews for each widget.

* Add a bit of padding in the previews struct.

* Add previews of the entries which have favorites.

* Mark the search bar as unredacted.

The widget selection menu on iOS will sometimes take a second or two to load, and will mark the widget as redacted while it does so.

This is useful for sections of the widget with dynamic content, but the search bar does not need to be included.