Configuration refresh improvements (#726)
* Add the background processing entitlement.

* Add a permitted identifier for the refresh task.

* Add functions to register and schedule tasks.

* Explicitly provide an ETag when making config requests.

* Attempt to refresh the configuration on every launch.

* Pass the ETag storage object to the request.

* Compile new content blocking rules on demand.

* Remove the existing rules list when updating.

* Post a notification to indicate that new tracker data is available.

* Ensure that the tracker data is up to date when compiling rules.

* Add an extra Pixel parameter to distinguish between fetch types.

* Remove an extra rule update which has no effect.

* Add a unit test for the APIHeader change.

* Replace the refresh interval added during debugging.

* Set a task expiration handler.

* Extract synchronous refresh logic into a new function.

This is so that it can be used from the background refresh task without using a UIApplication background task, which gives warnings at the 30 second mark.

* Call the compilation completion handler on the main thread.

* Fix a race condition with BackgroundTasks.

* Only submit tasks if there are none pending.

Otherwise the pending task will be replaced.

* Clean up a comment.

* Remove delegate changes in favor of an initializer.

The behaviour of reading the ETag is specific to the ContentBlockerRequest – mock implementations of the protocol shouldn’t have to take the storage object just to ignore it. For dependency injection that is specific to a single class, an initializer is more suitable.

* Fire a debug pixel when background task submission fails.

* Use the user defaults property wrapper in the config object.

* Rely on notifications for recompiling blocking rules.

* Check whether statistics need to be sent.

This is only checked if the config fetch class doesn’t need to refresh, implying that a refresh has already taken place recently, possibly in the background.

* Only send the task schedule failure pixel if on a physical device.

The simulator will cause this to throw with an unhelpful error message, we should avoid polluting the stats.