Fix potential memory leak in BloomFilterWrapper
As best I can tell, this is mostly theoretical
at this point---the BloomFilterWrapper is currently
retained for the lifetime of the program. If this
ever were *not* the case, though, we'd be leaking
the allocated BloomFilter object without a `delete`
corresponding to its `new`.

You can observe the leak in the test suite, which
*does* create temporary `BloomFilterWrapper`s: just
add a destructor to BloomFilter.hpp in the submodule
like this:

    ~BloomFilter() { printf("Destroyed Bloom\n"); }

...and observe that it's never called without this
patch, and consistently called with it.