Add conditional ActionMessageView width
In order to respect devices that have the width smaller than the maxWidth constant  I've added a conditional messageViewWidth that calculates if the window width is wider than maxWidth, if it is then we set a minimumHorizontalPadding to it.

I've also moved cornerRadius and windowBottomPadding to the Constants struct.

This fixes the issue: https://github.com/duckduckgo/iOS/issues/907