Set 'viewed' property of Tab model to false by default
The 'viewed' property being set as true by default is logically incorrect. This is why the jitter happens when adding new cell to tabs list/grid. In initial commit I worked around by setting 'viewed' to false right after new tab model is created and scheduling it being set to true after a delay. This was a quicker way as I wanted to show the before/after of the issue and get green light to post the PR.

But that way is not correct. Hence here the 'viewed' property is set to true when:
- HomeViewController's viewDidAppear method fires.
- TabViewController's viewDidAppear method fires.
- TabSwitcherViewController's  on select from touch and keyboard