Add String to silence iTC warning (#1030)
Note, this string will never be displayed to the user since we only use local speech recognizer and when local speech recognizer is used on iOS 15+ no alert permission dialog is displayed. This is being added to silence the "Missing Purpose String in Info.plist" warning when submitting to the app store.