Resolve the widget crash.
Users with a pre-folders install who have then installed a build with folders and have not yet migrated it will not have a top-level folder in Core Data.

This will be added the first time they launch a build that support folders, so an empty set of favorites is returned here to avoid the widget crashing.