Run PR Labels GHA workflow also on pushes to release branches (#1233)
Task/Issue URL: https://app.asana.com/0/414235014887631/1202594247175109/f

Description:
This makes release branches compliant with a branch protection rule that requires
a successful PR Labels check to be run on a commit before it's merged to develop.
Since labels field is not defined on push GitHub Actions event, the actual check inside the workflow
has to be skipped for pushes to a branch. It's still being triggered for pull requests.