Before processing result make sure speech recognition task is alive and not cancelled (#1318)
Task/Issue URL: https://app.asana.com/0/414709148257752/1202599139452865/f

Description:
Due to changes introduced in on-device speech recognition framework the result handler for SFSpeechRecognitionTask appears to be called even after it is cancelled. Those extraneous calls are with empty string result and thus override the query triggering loading of empty SERP.
To avoid this scenario we should check if SpeechRecognizer's internal state is coherent: that the speech recognition task is not cancelled and still assigned to a property before proceeding with processing the result. Otherwise the results are after reported after cancelling the task.