Force SwiftUI to refresh its views when objects change. (#12)
Task/Issue URL: https://app.asana.com/0/1203881841413756/1204121521899593/f
Tech Design URL:
CC:

Description:

This PR fixes tracker refreshing in the list view for AppTP. This turned out to be a SwiftUI problem: because we were passing the tracker objects straight to the view, SwiftUI's equality optimization was assuming the objects hadn't changed. To fix this, I'm passing the value types into the view, so SwiftUI will be able to see that their values have really changed.