Add a toggle between count and timestamp sorting options (#14)
Task/Issue URL: https://app.asana.com/0/1203881841413756/1204170442560315/f
Tech Design URL:
CC:

Description:

This PR adds a new segmented control that lets you sort trackers by count or timestamp. Count is the default.

Note that this does NOT persist between viewings of the screen. I could change this to use user defaults to persist the change if it's important, but I didn't bother so far since we already have a design project in place that may not include this feature.

This PR also adds a timestamp label to the tracker cell, but only in debug mode.