Fix `PrivacyIconView` Accessibility Traits Refresh (#1557)
Task/Issue URL: https://app.asana.com/0/414235014887631/1204297690570751/f
GH Issue URL: #1556

Description:
The PrivacyIconView is not refreshing the accessibility traits of the Dax Logo Image / Shield Button. This is mainly due to a typo on the willSet of the private(set) var icon: PrivacyIcon. Where the new value is not being sent to the method that updates the accessibility traits: updateAccessibilityLabels(for icon: PrivacyIcon)