Add inspectable webview support (#1703)
Task/Issue URL: https://app.asana.com/0/414709148257752/1204241755214739/f
Tech Design URL:
CC:

Description:

This PR adds support for the new inspectable flag on WKWebView. This allows internal users to turn on inspectability and gain access to the remote web inspector via Safari.

Note that this feature does not use the internal feature flag, it is tied exclusively to the Debug menu (which is itself internal only as far as end users are concerned) as the Debug menu can be accesssed when connected to Xcode even when you aren't technically in Internal mode.