Add Network Protection Settings item (#1836)
Task/Issue URL: https://app.asana.com/0/0/1205045707803897/f
Description:
As part of the work to test a barebones implementation of the PacketTunnelProvider providing Network Protection on iOS, we need to add an item in the main iOS Settings through which to access it. As this should currently only be available to internal users and also not even have code included in production builds, this is hidden behind both runtime and compile-time feature flags.