Specify iOS version as an argument to the `xcodebuild` run (#1890)
Task/Issue URL: https://app.asana.com/0/414709148257752/1205194825986261/f

Description:
GitHub updated the macOS 13 runner that defaults to iOS 17 beta simulators for Xcode 14.3.1 CI jobs when no version is explicitly provided. This results in slow build and unstable tests.
As a workaround we can specify an iOS version as an argument to the xcodebuild.