Point to BSK branch with NetP observer improvements (#1894)
Task/Issue URL: https://app.asana.com/0/0/1205210093476964/f
Tech Design URL:
CC:

Description:

For NetP Update Status View to match designs, I planned to rely on the shared Observer types to receive events from the PacketTunnnelProvider and other NetP types. However, on doing this, I noticed that we’re exposing the Subject type in order to synchronously access the most recent state.

Exposing these types as it means these events can be sent from anywhere and not just the source of the events. Rather than letting dependency on this propagate into iOS too, I decided to do a quick refactor to clean it up.

Additionally, I’ve created the associated mock types which I will also need throughout the iOS: Network Protection MVP TestFlight build for internal testing project. I decided now was the right time to create them as these types have been altered.

Creating the MockTunnelController is included in this task though it seems somewhat independent. This is so that it can be included in the same BSK release as it will also be needed for the parent task.

On iOS this is just a version bump for now to simplify the BSK release process.