NetP Status View Header (#1899)
Task/Issue URL: https://app.asana.com/0/0/1205084446087078/f

Description:

This is the first design implementation task in the overarching Network Protection Status View task and covers just the header view (that is, the views above the toggle view). I’m delaying on making this look good on iOS 14 for now as it will take some extra thought. But I am checking in with the Apple team and have created a follow-up task to fix iOS 14 at the end of the project if need be.