Status view error UI (#1915)
Task/Issue URL: https://app.asana.com/0/0/1205084446087081/f
Description:

This is the first PR for the Status View error states. The primary purpose of this is to get the design for review ASAP as this is the last task blocking design review. Therefore the main intention is to show an error.

This is initially achieved by just checking the ConnectionErrorObserver for a non-nil value, showing the generic connection failed error in that case, then showing no error in the case of a nil value.

A following PR will add handling of connection interruptions, however this will require some BSK changes so I didn’t want to block the design review to wait for that sometimes lengthy process.