Fix PacketTunnelProvider embedding for Alpha builds (#1932)
Task/Issue URL: https://app.asana.com/0/0/1205324333927521/f
Description:
On trying to distribute NetP through the Alpha channel, the build were being rejected by App Store Connect as their was an invalid file containing the word app. This turned out to be because the simple cp -r I was doing to move the extension was in fact not moving the extension for Archive builds, but an alias.

I had another look at the build output of Xcode’s Embed App Extensions Build Phase which uses an Xcode-internal tool builtin-copy. rsync achieves the same result and has broadly the same options. The --copy-links option is needed to resolve symlinked .appex bundles.