Separate bundle and group ids in iOS Alpha config file (#1957)
Task/Issue URL: https://app.asana.com/0/0/1205101983701300/f

Description:

On implementing PR: Separate app groups across Alpha and other builds, the app groups were updated for Alpha builds, but the generic Configuration.xcconfig that was used for all build configs has defined a single app group prefix for all configurations. This resulted in a crash on startup.

To fix this, we’ll need to define a separate app group prefix. This could be done through an override in the Xcode Build Settings, but I feel it’s cleaner to do this through a Configuration-Alpha.xcconfig.