Add nightly CodeQL workflow (#1958)
Task/Issue URL: https://app.asana.com/0/0/1205360222695977/f
Tech Design URL:
CC: @ayoy

Description:

This PR adds a CodeQL workflow which runs nightly. I've chosen not to run it on PR commits as the job can be quite slow (up to 30 mins in some tests), but I'm open to changing when/where this runs.

Changes will be drafted this week to propose how the alerts are handled, but for now we can begin to let the job run and see how it behaves over time.