Fix address bar updates from occurring from download paths with no document.
When we receive a 302 redirect to a file download path, we inadvertently set the address bar to the origin of the download path while leaving a stale document in the renderer. This can lead to address bar spoofing, so instead we set the address bar to about:blank while the download is in progress.