Update VPN active user check and debug options (#2269)
Task/Issue URL: https://app.asana.com/0/0/1206165514403612/f
Tech Design URL:
CC:

Description:

This PR updates the active user pixel check to accurately reflect the conditions under which you need to meet to be counted as a potential user. The old check was looking for the active flag only, which isn't accurate as it includes non-US users.