Reenable toggle on disallowing vpn (#2404) (#2416)
Task/Issue URL: https://app.asana.com/0/0/1206463994034099/f

Description:

When “Don’t Allow” is tapped, the dialog is dismissed and the toggle is greyed out. We have to back out of the screen and reopen the VPN screen to give it another attempt.

Reenable the button again when the VPN permission is denied.

This can be solved by simply reenabling the toggle after it’s been disabled for 2 seconds, something we were already sort of doing to stop spamming. This just slightly reworks the logic to cover more scenarios.