Add debug command to disable VPN (#2391)
Task/Issue URL: https://app.asana.com/0/414235014887631/1206448305134854/f
Tech Design URL:
CC:

Description:

This PR adds a "Disable VPN" command to the debug menu, which can disable Connect on Demand and disable the VPN from within the extension. This is so that we can test that this iOS 17 change really works, before beginning to rely on it.