Subscriptions: Fix thread issue on Subscription Restore (#2719)
ask/Issue URL: https://app.asana.com/0/1204099484721401/1207054737385929/f

Description:
There is a threading issue on deInit for the SubscriptionFlow and SubscriptionRestore Viewmodels causing simultaneous calls to cleanup() in the SubFeature. (While I was unable to reproduce the issue, it’s there.)

Since subFeature is now a single instance everywhere, this adds a new ViewModel that holds and cleans the subFeature object instead of delegating the tasks to the internal viewModels.