Replace deprecated tab preview renderer (#2814)
Task/Issue URL: https://app.asana.com/0/414235014887631/1207228085347616/f
Tech Design URL:
CC:

Description:

This PR replaces usage of UIGraphicsBeginImageContextWithOptions with UIGraphicsImageRenderer.

UIGraphicsBeginImageContextWithOptions was deprecated in iOS 17, and is reported to be unstable. We're seeing crashes with it ourselves, and the assertion suggests migrating to the new API, so here we are.