Clean up subscription state caching in settings logic (#2925)
Task/Issue URL: https://app.asana.com/0/414709148257752/1207502776093446/f

Description:
To reduce the amount of network traffic when opening settings the subscription section state is being cached for immediate view state initialisation and then various state properties are being refreshed. Inspecting the code it turns out that in some of the scenarios of for example connectivity failure, cache expiry and/or access token missing the cached state may be outdated or wrong.