Fix search bar selection when coming back to the app (#3065)
Task/Issue URL: https://app.asana.com/0/414709148257752/1207736781372157/f
Tech Design URL:
CC:

Description:

Fix Selection when on the serp anc coming back to the app.

Steps to test this PR:

Press fire button.
Ensure the URL bar is selected and keyboard is up.
Leave the app for 6s.
Come back to the app - search for something.
Tap on search bar - text should not be selected, cursor should be at the end.
Hide keyboard, then background the app for 6s.
Come back to the app -> tap on search bar. Whole text should be selected.
Dismiss keyboard, tap on search again. Text should not be selected.
Navigate to non-serp page.
Tap on the url - it should be selected.