Allow activating subscription for internal users via debug menu (#3117)
Task/Issue URL: https://app.asana.com/0/72649045549333/1207811184536020/f

Description:

Add to debug menu option to start the subscription activation flow. This option is to allow access to the flow even when the purchase is disallowed in given region (App Store build for non-US user).

Additional changes are renaming and cleanup of obsolete items in the "Subscription" debug menu.