Show Intro message for existing users on New Tab Page (#3173)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.

⚠️ If you're an external contributor, please file an issue first before
working on a PR, as we can't guarantee that we will accept your changes
if they haven't been discussed ahead of time. Thanks!
-->

Task/Issue URL:
https://app.asana.com/0/1206226850447395/1207951131408752/f
Tech Design URL:
CC:

**Description**:

Adds a message for existing users about redesign of New Tab Page. I'm
using the presence of install date during app launch to determine if the
user is new.
Existing implementation also checks whether the feature flag is set to
be releasable remotely and does nothing if it is not. Altering the code
is necessary to verify final behavior.

**Steps to test this PR**:
1. Install and run any other build (without the intro message). If you
have any other build run already you can skip this step.
1. Change `NewTabPageManager.isAvailableInPublicRelease` to return
`true`, to override setup restrictions. Run build from this branch.
1. Set _internal user state_ and enable New Tab Page Sections feature
flag in Debug menu.
1. Intro message should be visible on new tab page as well as for next
two new tab page impressions.
1. Tapping on X button should hide message forever.
1. Remove app.
1. Launch and enable the feature in Debug menu again - intro message
should not be visible.

**Finally:** With code unchanged message should not initialize at all.
It can be reset and verified in Debug menu.