Break DuckPlayer ref cycle (#3206)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.

⚠️ If you're an external contributor, please file an issue first before
working on a PR, as we can't guarantee that we will accept your changes
if they haven't been discussed ahead of time. Thanks!
-->

Task/Issue URL:
https://app.asana.com/0/414709148257752/1207998631270910/f

**Description**:
Breaks reference cycle between DuckPlayer and TabViewController, which
can contribute to containers not being properly cleaned up

**Steps to test this PR**:
1.  Fire up the app, navigate away
2. Check Memory Graph, confirm the duckPlayer property in
TabViewController is now a weak reference,
3. Fire button everything.  
4. There should not be leftovers.