Fixes DuckPlayer Experiment issue (#3421)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.

⚠️ If you're an external contributor, please file an issue first before
working on a PR, as we can't guarantee that we will accept your changes
if they haven't been discussed ahead of time. Thanks!
-->

Task/Issue URL:
https://app.asana.com/0/1207252092703676/1208450923559111/f

**Description**:
- Fix an issue with DP experiment that causes Control users to not see
overlays