Do not notify the FE on experiment activation (#3450)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.

⚠️ If you're an external contributor, please file an issue first before
working on a PR, as we can't guarantee that we will accept your changes
if they haven't been discussed ahead of time. Thanks!
-->

Task/Issue URL:
https://app.asana.com/0/1204099484721401/1208531438887018/f

**Description**:
- Mitigates Frontend issue causing DuckPlayer to be stuck in Youtube
Mode