Blind attempt to fix Omnibar-related crash (#3500)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.

⚠️ If you're an external contributor, please file an issue first before
working on a PR, as we can't guarantee that we will accept your changes
if they haven't been discussed ahead of time. Thanks!
-->

Task/Issue URL:
https://app.asana.com/0/414709148257752/1208653357224780/f

**Description**:
Attempt to fix crash:
Sentry:
https://errors.duckduckgo.com/organizations/ddg/issues/169611/?project=8&query=is%3Aunresolved+firstSeen%3A-15d&referrer=issue-stream&sort=freq&statsPeriod=14d&stream_index=0