adding impression pixels for duckplayer in landscape mode (#3493)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.

⚠️ If you're an external contributor, please file an issue first before
working on a PR, as we can't guarantee that we will accept your changes
if they haven't been discussed ahead of time. Thanks!
-->

Task/Issue URL:
https://app.asana.com/0/1201141132935289/1208637202774549/f
Tech Design URL:
CC:

**Description**:

the DuckPlayer frontend will start sending a generic 'telemetry' events
soon - this PR adds support for the first one, but in a format that can
be expanded later without needing new handlers.