Remove NewTabPage retain cycles (#3532)
Task/Issue URL:
https://app.asana.com/0/1206226850447395/1208686031091434/f
Tech Design URL:
CC:

**Description**:

Leak 1:
`NewTabPageViewController` was not dismissed properly, this caused it to
stay in memory as a child view controller of `MainViewController`.
In effect it was possible to dismiss FaviconFetcherTutorial. Removing
the leak required to do additional changes to make it work. I moved it
to `MainViewController`, so that it's not dependent on NewTabPage.

Leak 2:
`NewTabPageSettingsModel` was leaking via strong reference present in
`NTPSettingItem`.

**Steps to test this PR**:
1. Set up Sync.
2. Add favorite.
3. On another synced device open New Tab Page
4. Favicon Tutorial should appear, see if buttons work, dismissing the
tutorial
5. Open and close New Tab Page a few times.
6. Open Memory Graph Debugger, verify only single instance is present
for `NewTabPageViewController` and `NewTabPageSettingsModel*` objects.

**Definition of Done (Internal Only)**:

* [ ] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

---
###### Internal references:
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)