Fix autofill search authentication looping (#3598)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.

⚠️ If you're an external contributor, please file an issue first before
working on a PR, as we can't guarantee that we will accept your changes
if they haven't been discussed ahead of time. Thanks!
-->

Task/Issue URL:
https://app.asana.com/0/1203822806345703/1208755477452311/f
Tech Design URL:
CC:

**Description**:
Fixes an issue where if the Passwords screen is dismissed while in
search mode, UISearchController prevented the Passwords screen from
being released from memory. I’ve also found and fixed a memory leak with
the autofill breakage reporter.