Re-prompt for crash reporting (#3595)
Task/Issue URL:
https://app.asana.com/0/1208592102886666/1208630312625695/f
Tech Design URL:
https://app.asana.com/0/1208592102886666/1208660326715650/f

**Description**:
- Introduce a feature flag for use with this project overall, currently
set to internal only
- Introduce a new UserDefaults value
(```crashCollectionShouldRevertOptedInStatusTrigger```) that can be used
to force the client to revert the existing
```crashCollectionOptInStatus``` to .unknown, only if the value is
currently .optedIn. This new trigger is implemented as an Int, so that
we can ensure the user’s Opt-In state will be reverted only once after
the feature is turned on. In the future if we need this functionality
again, we’d simply bump the garget value and the opt-in state would be
reset again.

See asana task for specific requirements, and the parent project for
more context if necessary.