Update Fastfile to wrap lanes in a platform block (#3711)
Task/Issue URL: https://app.asana.com/0/0/1208949587447312/f
Tech Design URL:
CC:

Description:

This PR updates the iOS Fastfile to wrap lanes in a platform block.

This introduces zero changes to the usability of the Fastfile as it is only intended to lay the groundwork for the move to the monorepo. However, the diff is noisy since wrapping all lanes in a block caused an indentation change. Sorry!