Crash report cohort ID support for iOS (#3692)
Task/Issue URL:
https://app.asana.com/0/1208592102886666/1208759541597499/f
Tech Design URL:
https://app.asana.com/0/1208592102886666/1208660326715650/f

**Description**:
DO NOT MERGE - this is a draft for input, not ready to go live yet.

iOS client support for CRCID send/receive (primarily supported in BSK,
with changes under review in [BSK
#1116](https://github.com/duckduckgo/BrowserServicesKit/pull/1116)).
This is pretty straightforward, just conforming to CrashCollection’s new
init signature, and clearing CRCIDs when the user opts out of crash
reporting. BSK handles everything else.

**Steps to test this PR**:
Note: Must be tested on a physical device, as the simulator does not
produce crash logs (and thus doesn’t find and upload them either).

To cause and report a crash:
1. Launch the app and force a crash, which can be done from Settings →
All Debug Options → Crash (fatal error) or similar. Note that Crash
(CPU/Memory) does not appear to produce a crash log, and thus won’t
trigger crash uploading.
2. Launch the app again (easiest with a debugger)
1. For the first crash of an app install: You will be prompted to opt in
or out of crash reporting when the app is launched. Opt in and watch
logs for “crcid” and you should see logs from CrashReportSender:56, and
CrashCollection:95-109.
2. On subsequent crashes, when opted in, you should see statements
confirming the received crcid was sent, and that the server returned
either the same matching one, or a new one (in which case the new one
should be stored and used on subsequent crash reports)

To test clearing of the crcid when opting out:
1. Navigate to Settings → About and switch “Send Crash Reports” off,
then back on again (this step should clear the crcid)
2. Follow steps from “To cause and report a crash” above, and confirm
that the crash is submitted without an initial crcid, and that the
server assigns one and it is stored (causing and uploading a second
crash should confirm this new value is used on send).