Switch to using maestro action (#3783)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.

⚠️ If you're an external contributor, please file an issue first before
working on a PR, as we can't guarantee that we will accept your changes
if they haven't been discussed ahead of time. Thanks!
-->

Task/Issue URL:
https://app.asana.com/0/1204165176092271/1209158713404043/f
Tech Design URL:
CC:

**Description**:
Maestro CLI fail quite often because it drops its connection to Robin
CI. The result is that we get failures in GHA for tests that actually
complete successfully in Robin.

This PR updates our GHA workflow to use Maestro Cloud Action that
doesn't seem to have this issue.

The problem with the official Maestro Action is that it doesn't consider
timeouts or other infrastructure failures as failures and marks tests
that fall into that bucket as successful. This is [known
problem](https://github.com/mobile-dev-inc/action-maestro-cloud/issues/28)
that they haven't solved yet. To mitigate this, I've forked the action
and introduced a fix for the most prominent issue related to timeouts.
I plan to open a PR with these fixes against the official repository
asap. In the meantime, this PR uses my fork.

<!--
If at any point it isn't actively being worked on/ready for
review/otherwise moving forward strongly consider closing it (or not
opening it in the first place). If you decide not to close it, use Draft
PR while work is still in progress or use `DO NOT MERGE` label to
clarify the PRs state and comment with more information.
-->

**Steps to test this PR**:
1. Run end-to-end tests on CI from this branch and validate they are
green (or you can check the result of a recent run from this branch
here: https://github.com/duckduckgo/iOS/actions/runs/12786928298)
2.

<!--
Before submitting a PR, please ensure you have tested the combinations
you expect the reviewer to test, then delete configurations you *know*
do not need explicit testing.

Using a simulator where a physical device is unavailable is acceptable.
-->

**Definition of Done (Internal Only)**:

* [x] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

**Copy Testing**:

* [ ] Use of correct apostrophes in new copy, ie `’` rather than `'`

**Orientation Testing**:

* [ ] Portrait
* [ ] Landscape

**Device Testing**:

* [ ] iPhone SE (1st Gen)
* [ ] iPhone 8
* [ ] iPhone X
* [ ] iPhone 14 Pro
* [ ] iPad

**OS Testing**:

* [ ] iOS 15
* [ ] iOS 16
* [ ] iOS 17

**Theme Testing**:

* [ ] Light theme
* [ ] Dark theme

---
###### Internal references:
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)